+++ code/src/cljs/sixsq/nuvla/ui/dashboard/subs.cljs
@@ -59,4 +59,5 @@
     (let [deployment-params-by-name (->> (get deployments-params-map id)
                                          (map (juxt :name identity))
                                          (into {}))]
-      (utils/resolve-url-pattern url-pattern deployment-params-by-name))))
+      (when (utils/running-replica? deployment-params-by-name)
+        (utils/resolve-url-pattern url-pattern deployment-params-by-name)))))