+++ code/src/cljs/sixsq/nuvla/ui/dashboard_detail/subs.cljs
@@ -46,12 +46,14 @@
 (reg-sub
   ::deployment-parameters
   (fn [db]
-    (::spec/deployment-parameters db)))
-
+    (->> db
+         ::spec/deployment-parameters
+         (sort-by :name))))
 
 
 (reg-sub
   ::url
   :<- [::deployment-parameters]
   (fn [deployment-parameters [_ url-pattern]]
-    (dashboard-utils/resolve-url-pattern url-pattern deployment-parameters)))
+    (when (dashboard-utils/running-replica? deployment-parameters)
+      (dashboard-utils/resolve-url-pattern url-pattern deployment-parameters))))